+++ CHANGELOG.md
@@ -24,6 +24,11 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 ---
 
+### Changed
+
+-   [#73](https://github.com/FC4E-CAT/fc4e-cat-api/pull/73)    -  CAT-159 Change assessment id from bigint to uuid.
+
+
 ## 1.1.0 - 2023-09-01
 
 ---